requirements.txt
aniso8601==3.0.2
beautifulsoup4==4.7.1
unidecode
certifi==2019.3.9
chardet==3.0.4
Cython==0.29.7
enum34==1.1.6
fasttext==0.8.3
future==0.17.1
googlemaps==3.0.2
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
nltk==3.4.1
numpy==1.16.3
oauthlib==3.0.1
pbr==5.2.0
pkg-resources==0.0.0
promise==2.2.1
PySocks==1.6.8
python-google-places==1.4.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
Rx==1.6.1
scipy==1.2.1
six==1.12.0
soupsieve==1.9.1
stevedore==1.30.1
tweepy==3.7.0
urllib3==1.24.3
